==> notebooks/02 Genius Scraper.ipynb <==
# **02 Genius Scraper**


### Using Genius to scrape data on the title, artist and song lyrics

---

### **Set-up ⚙️**

Import necessary packages

*⚠️ Note: Do not run this more than once. Restart the kernel before running this code chunk.*

import json
import pandas as pd
from tqdm import tqdm
import os
os.chdir(os.path.expanduser("../"))                 # change directory to main project directory

from functions.genius_functions import *            # imports custom functions for genius scraping

Check that we are in the correct current working directory

*⚠️ Note: We should be in the main project directory*

print("Current working directory:", os.getcwd())

Open JSON file containing credentials

*⚠️ Note: Our credentials should be stored in a file titled `credentials.json` and stored in the root of the project folder*

credentials_file_path = './credentials.json'

with open(credentials_file_path, 'r') as f:
    credentials = json.load(f)

Initialise a new session

my_session = requests.Session()

---

### **Data Scraping 🔍**

Import cleaned YouTube data into a dataframe

cleaned_df = pd.read_csv('./data/cleaned_youtube_final_data.csv', index_col=0)

Using the `search_genius` function, scrape data on the song title, artist and Genius URL using each video title as the query parameter, and store the scraped data into a new dataframe

*⚠️ Note: This code chunk will take a long time to run and complete*
<br>*⚠️ Note: Genius API access token should be saved under the key `genius_client_access_token` in the `credentials.json` file*

scraped_df = pd.DataFrame([search_genius(q, credentials['genius_client_access_token']) for q in tqdm(cleaned_df['video_title'])])

Using the `scrape_lyrics` function, scrape data on the song lyrics of each song and add the scraped data into the existing dataframe

*⚠️ Note: This code chunk will take a long time to run and complete*

scraped_df['lyrics'] = scraped_df.apply(lambda row: scrape_lyrics(my_session, row['URL']) if row['URL'] else '', axis=1)

Save dataframe to csv

scraped_df.to_csv('./data/raw_data_with_lyrics.csv')

---

### **Data Cleaning 🧹**

Merge dataframes of data scraped from YouTube and Genius into one large dataframe

# load dataframes from csv files
youtube_df = pd.read_csv('./data/cleaned_youtube_final_data.csv')
genius_df = pd.read_csv('./data/raw_data_with_lyrics.csv')

large_df = genius_df.merge(youtube_df)

# drop unnecessary columns
large_df = large_df.drop(columns=['Unnamed: 0', 'video_title'])

# rename columns for standardisation
large_df = large_df.rename(columns=str.lower)

Save final dataframe as a CSV file

large_df.to_csv('./data/raw_compiled_data.csv', index=False)
